// console/src/generated/RuleDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RuleConfigDto } from "./RuleConfigDto";

export type RuleDto = { id: number, rule_set_id: number, config: RuleConfigDto, filter_model_pattern: string | null, filter_operations: Array<string> | null, filter_header_pattern: string | null, sort_order: number, enabled: boolean, };